exfat_mkfs: define exfat_mkfs to make exfat file system
we can reuse this code at other component
